Restructure, add rules-to-vars util
