+++ packages/themes/src/composables/index.ts
@@ -1,0 +1,2 @@
+export * from './use-fonts.js';
+export * from './use-theme.js';
