+++ packages/themes/src/index.ts
@@ -1,4 +1,5 @@
-export * from './schema.js';
-export * from './store.js';
-export * from './use-theme.js';
-export { default as ThemeProvider } from './theme-provider.vue';
+export * from './components/index.js';
+export * from './composables/index.js';
+export * from './schemas/index.js';
+export * from './stores/index.js';
+export * from './utils/index.js';
